refactor: split looks, node-utils, operators, expressions, control-flow, block-utils
Extract validation helpers from looks.js, type checks from node-utils.js,
math operators from operators.js, literal visitors from expressions.js,
def visitor from control-flow.js, and block creation from block-utils.js.

Refs #357
