+++ packages/scratch-gui/src/lib/ruby-to-blocks-converter/operators.js
@@ -1,43 +1,14 @@
 import _ from 'lodash';
 import {RubyToBlocksConverterError} from './errors';
-
-const Math = '::Math';
-const MathE = '::Math::E';
+import OperatorsMath from './operators-math';
 
 /**
  * Operators converter
  */
 const OperatorsConverter = {
     register: function (converter) {
-        // String-typed variable + any: use operator_join (must be registered before the numeric + handler)
-        converter.registerOnSend('variable', '+', 1, params => {
-            const {receiver, args} = params;
-            const variable = converter.lookupVariableFromVariableBlock(receiver);
-            if (!variable || variable.dataType !== 'string') return null;
-
-            let rh = args[0];
-            if (_.isArray(rh)) {
-                if (rh.length !== 1) return null;
-                rh = rh[0];
-            }
-
-            // Boolean values are not valid operands for string concatenation
-            if (converter._isTrue(rh) || converter._isFalse(rh)) return null;
-
-            if (!converter._isStringOrBlock(rh)) return null;
-
-            const block = converter._createBlock('operator_join', 'value');
-            converter._addTextInput(block, 'STRING1', receiver, 'apple');
-            converter._addTextInput(block, 'STRING2', converter._isNumber(rh) ? rh.toString() : rh, 'banana');
-            return block;
-        });
-
-        converter.registerOnSend('self', 'rand', 1, params => {
-            const {args} = params;
-            if (!converter._isBlock(args[0]) || args[0].opcode !== 'ruby_range') return null;
-
-            return converter._changeBlock(args[0], 'operator_random', 'value');
-        });
+        // Register arithmetic and math function handlers
+        OperatorsMath.registerMath(converter);
 
         converter.registerOnSend(['string', 'block', 'variable'], '[]', 1, params => {
             const {receiver, args} = params;
@@ -93,46 +64,6 @@ const OperatorsConverter = {
             converter._addTextInput(block, 'STRING1', receiver, 'apple');
             converter._addTextInput(block, 'STRING2', args[0], 'a');
             return block;
-        });
-
-        ['+', '-', '*', '/', '%'].forEach(operator => {
-            converter.registerOnSend(['variable', 'number', 'block'], operator, 1, params => {
-                const {receiver, args} = params;
-                let rh = args[0];
-                if (_.isArray(rh)) {
-                    if (rh.length !== 1) return null;
-                    rh = rh[0];
-                }
-
-                // Skip string-typed variables for + (handled by the dedicated string handler)
-                if (operator === '+' && converter.isVariableBlock(receiver)) {
-                    const variable = converter.lookupVariableFromVariableBlock(receiver);
-                    if (variable && variable.dataType === 'string') return null;
-                }
-
-                // Boolean values are not valid operands for numeric arithmetic
-                if (converter._isTrue(rh) || converter._isFalse(rh)) return null;
-
-                if (!converter._isNumberOrBlock(rh)) return null;
-
-                let opcode;
-                if (operator === '+') {
-                    opcode = 'operator_add';
-                } else if (operator === '-') {
-                    opcode = 'operator_subtract';
-                } else if (operator === '*') {
-                    opcode = 'operator_multiply';
-                } else if (operator === '/') {
-                    opcode = 'operator_divide';
-                } else {
-                    opcode = 'operator_mod';
-                }
-
-                const block = converter._createBlock(opcode, 'value');
-                converter._addNumberInput(block, 'NUM1', 'math_number', receiver, '');
-                converter._addNumberInput(block, 'NUM2', 'math_number', rh, '');
-                return block;
-            });
         });
 
         converter.registerOnSend(['string', 'block'], '+', 1, params => {
@@ -404,14 +335,6 @@ const OperatorsConverter = {
             return block;
         });
 
-        converter.registerOnSend(['variable', 'number', 'block'], 'round', 0, params => {
-            const {receiver} = params;
-
-            const block = converter._createBlock('operator_round', 'value');
-            converter._addNumberInput(block, 'NUM', 'math_number', receiver, '');
-            return block;
-        });
-
         converter.registerOnSend('symbol', 'to_s', 0, params => {
             const {receiver} = params;
             const symbolName = converter._getSymbolValue(receiver);
@@ -450,83 +373,6 @@ const OperatorsConverter = {
             }
             converter._addNumberInput(block, 'NUM2', 'math_number', 0, '');
             block.comment = converter._createComment('@ruby:method:to_i', block.id);
-            return block;
-        });
-
-        ['abs', 'floor', 'ceil'].forEach(methodName => {
-            converter.registerOnSend(['variable', 'number', 'block'], methodName, 0, params => {
-                const {receiver} = params;
-
-                let operator = methodName;
-                if (methodName === 'ceil') {
-                    operator = 'ceiling';
-                }
-                const block = converter._createBlock('operator_mathop', 'value');
-                converter._addField(block, 'OPERATOR', operator);
-                converter._addNumberInput(block, 'NUM', 'math_number', receiver, '');
-                return block;
-            });
-        });
-
-        ['sqrt', 'sin', 'cos', 'tan', 'asin', 'acos', 'atan', 'log', 'log10'].forEach(methodName => {
-            converter.registerOnSend(Math, methodName, 1, params => {
-                const {args} = params;
-                let rh = args[0];
-                if (_.isArray(rh)) {
-                    if (rh.length !== 1) return null;
-                    rh = rh[0];
-                }
-
-                if (!converter._isNumberOrBlock(rh)) return null;
-
-                let operator;
-                switch (methodName) {
-                case 'log':
-                    operator = 'ln';
-                    break;
-                case 'log10':
-                    operator = 'log';
-                    break;
-                default:
-                    operator = methodName;
-                }
-                const block = converter._createBlock('operator_mathop', 'value');
-                converter._addField(block, 'OPERATOR', operator);
-                converter._addNumberInput(block, 'NUM', 'math_number', rh, '');
-                return block;
-            });
-        });
-
-        converter.registerOnSend(MathE, '**', 1, params => {
-            const {args} = params;
-            let rh = args[0];
-            if (_.isArray(rh)) {
-                if (rh.length !== 1) return null;
-                rh = rh[0];
-            }
-
-            if (!converter._isNumberOrBlock(rh)) return null;
-
-            const block = converter._createBlock('operator_mathop', 'value');
-            converter._addField(block, 'OPERATOR', 'e ^');
-            converter._addNumberInput(block, 'NUM', 'math_number', rh, '');
-            return block;
-        });
-
-        converter.registerOnSend('number', '**', 1, params => {
-            const {receiver, args} = params;
-            let rh = args[0];
-            if (_.isArray(rh)) {
-                if (rh.length !== 1) return null;
-                rh = rh[0];
-            }
-
-            if (!receiver.value === 10) return null;
-            if (!converter._isNumberOrBlock(rh)) return null;
-
-            const block = converter._createBlock('operator_mathop', 'value');
-            converter._addField(block, 'OPERATOR', '10 ^');
-            converter._addNumberInput(block, 'NUM', 'math_number', rh, '');
             return block;
         });
